backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.42
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.10.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
email-validator==2.2.0
httpx==0.28.1
beautifulsoup4==4.13.4
lxml==6.0.0
curl_cffi==0.14.0
browser-cookie3==0.20.1
